using uint64_t for handles and size_t for size

+++ include/ne_renderer_pass.h
@@ -66,7 +66,7 @@ void ne_render_pass_set_pipeline(NERenderPass *pass, NEPipelineHandle pipeline);
  *              `NEPipelineDesc::vertex_layouts`).
  * - `buffer` : Buffer handle (must have NE_BUFFER_USAGE_VERTEX).
  */
-void ne_render_pass_set_vertex_buffer(NERenderPass *pass, uint32_t slot,
+void ne_render_pass_set_vertex_buffer(NERenderPass *pass, uint64_t slot,
                                       NEBufferHandle buffer);
 
 /**
@@ -97,7 +97,7 @@ void ne_render_pass_set_index_buffer(NERenderPass *pass, NEBufferHandle buffer,
  * - `size`  : Size in bytes (must not exceed backend limit; typically 128-256 B).
  */
 void ne_render_pass_set_uniform_data(NERenderPass *pass, NEShaderStage stage,
-                                     uint32_t slot, const void *data, uint32_t size);
+                                     uint64_t slot, const void *data, size_t size);
 
 /**
  * Issue a non-indexed draw call.
@@ -106,8 +106,8 @@ void ne_render_pass_set_uniform_data(NERenderPass *pass, NEShaderStage stage,
  * - `first_vertex` : Offset of the first vertex.
  * - `vertex_count` : Number of vertices to draw.
  */
-void ne_render_pass_draw(NERenderPass *pass, uint32_t first_vertex,
-                         uint32_t vertex_count);
+void ne_render_pass_draw(NERenderPass *pass, uint64_t first_vertex,
+                         uint64_t vertex_count);
 
 /**
  * Issue an indexed draw call.
@@ -117,8 +117,8 @@ void ne_render_pass_draw(NERenderPass *pass, uint32_t first_vertex,
  * - `first_index`   : Offset into the index buffer (in elements, not bytes).
  * - `vertex_offset` : Value added to each index before fetching the vertex.
  */
-void ne_render_pass_draw_indexed(NERenderPass *pass, uint32_t index_count,
-                                 uint32_t first_index, int32_t vertex_offset);
+void ne_render_pass_draw_indexed(NERenderPass *pass, uint64_t index_count,
+                                 uint64_t first_index, int64_t vertex_offset);
 
 /* ======================================================================== */
 /* Compute pass commands                                                    */
@@ -136,7 +136,7 @@ void ne_compute_pass_set_pipeline(NEComputePass *pass, NEComputePipelineHandle p
  * - `slot`   : Binding slot / index.
  * - `buffer` : Buffer handle (must have NE_BUFFER_USAGE_STORAGE).
  */
-void ne_compute_pass_set_storage_buffer(NEComputePass *pass, uint32_t slot,
+void ne_compute_pass_set_storage_buffer(NEComputePass *pass, uint64_t slot,
                                         NEBufferHandle buffer);
 
 /**
@@ -145,7 +145,7 @@ void ne_compute_pass_set_storage_buffer(NEComputePass *pass, uint32_t slot,
  * Same semantics as `ne_render_pass_set_uniform_data` but targets the
  * compute stage.
  */
-void ne_compute_pass_set_uniform_data(NEComputePass *pass, uint32_t slot,
+void ne_compute_pass_set_uniform_data(NEComputePass *pass, uint64_t slot,
                                       const void *data, uint32_t size);
 
 /**
@@ -154,8 +154,8 @@ void ne_compute_pass_set_uniform_data(NEComputePass *pass, uint32_t slot,
  * Parameters:
  * - `group_count_x/y/z` : Number of workgroups in each dimension.
  */
-void ne_compute_pass_dispatch(NEComputePass *pass, uint32_t group_count_x,
-                              uint32_t group_count_y, uint32_t group_count_z);
+void ne_compute_pass_dispatch(NEComputePass *pass, uint64_t group_count_x,
+                              uint64_t group_count_y, uint64_t group_count_z);
 
 #ifdef __cplusplus
 }
